requirements.txt
Flask
aiohttp
gunicorn
requests
Pillow